Merging items seems to be working
Testing this will be tricky. The error messages remain in
Map::mergeItems while Im not certain it works well.

Decided not to refactor MapItems to include Cells as this is reduntant
information. Instead, when it needs to check points, map will pass a
reference to that list.

Note: Change TileType to include a flag called 'terminator' to make it
easy to know when roads are complete, and remove 'connected' as it wont
be necessary.

TODO: Documentation. MapItem, players need to be associated to a
tile/TypeIdentifier to display it in the GUI.