Modified main loop state-machine
Added PlayerManager to manage turns and etc. The main will be basically this class dancing with Map.

Next, work on Map::selectItemAt. This should determine which MapItem the player selected and call Map::setPiece.

Also, Map::render needs to find out what is the state (from the last_played pointer) and display which MapItems are unclaimed.